feat(core): 完成第四阶段模块拆分 - message_builder 和 version_checker
新增模块：
- core/message_builder.py (~280行): 消息构建和状态显示
- core/version_checker.py (~200行): 版本检查和更新通知

主要变更：
- 从 main.py 提取消息构建相关方法到 MessageBuilder 类
- 从 main.py 提取版本检查相关方法到 VersionChecker 类
- 更新 main.py 中的方法调用使用新模块
- 删除 main.py 中已迁移的旧方法
