+++ core/__init__.py
@@ -7,8 +7,19 @@
 from .config_manager import ConfigManager
 from .limiter import Limiter
 from .logger import Logger
+from .message_builder import MessageBuilder
 from .redis_client import RedisClient
 from .security import Security
 from .usage_tracker import UsageTracker
+from .version_checker import VersionChecker
 
-__all__ = ["Logger", "RedisClient", "ConfigManager", "Limiter", "Security", "UsageTracker"]
+__all__ = [
+    "Logger",
+    "RedisClient",
+    "ConfigManager",
+    "Limiter",
+    "Security",
+    "UsageTracker",
+    "MessageBuilder",
+    "VersionChecker",
+]
